UPD - Página principal
Componentizado o interior da página para melhor organização do código correspondente aos menus selecionados.

Consequentemento realizada a criação de 5 novos componentes:
-HomeInner
-DownloadsInner
-ComunidadeInner
-SobreInner
-ContatoInner
