+++ src/app/components/ComunidadeInner/ComunidadeInner.module.css
@@ -1,0 +1,6 @@
+.containerComunidade {
+    display: grid;
+    justify-content: center;
+    min-height: 60vh;
+    align-items: center;
+}
