+++ src/app/components/ComunidadeInner/index.tsx
@@ -1,0 +1,11 @@
+import { useState } from "react";
+import styles from "./ComunidadeInner.module.css";
+
+export default function ComunidadeInner() {
+  
+  return (
+    <div className={styles.containerComunidade}>
+        <h1>Comunidade</h1>
+    </div>
+  );
+}
