+++ src/app/components/ContatoInner/ContatoInner.module.css
@@ -1,0 +1,33 @@
+.containerContato{
+    background-image: url("../../../../public/bgContato.png");
+    width: 100%;
+    height: 100%;
+    background-position: center;
+    background-repeat: no-repeat;
+    background-size: cover;
+    display: flex;
+    padding: 3rem 4rem;
+}
+
+.containerForm{
+    background-color: #fff;
+    width: 50%;
+    height: 100%;
+    border-radius: 8px;
+    -webkit-box-shadow: 0px 2px 16px 1px rgba(0,0,0,0.75);
+    -moz-box-shadow: 0px 2px 16px 1px rgba(0,0,0,0.75);
+    box-shadow: 0px 2px 16px 1px rgba(0,0,0,0.75);
+    padding: 2rem;
+    text-align: center;
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+}
+
+.imgContato{
+    display: flex;
+    position: relative;
+    bottom: 20px;
+    left: 70px;
+    transform: rotate(-14deg);
+}
