+++ src/app/components/HomeInner/HomeInner.module.css
@@ -1,0 +1,11 @@
+.containerHome {
+    width: 100%;
+    min-width: 100%;
+    display: flex;
+    height: 100%;
+    min-height: 70vh;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+    gap: 1rem;
+  }
