+++ src/app/components/SobreInner/SobreInner.module.css
@@ -1,0 +1,45 @@
+.containerSobre {
+    background-image: url("../../../../public/fundoSobre.png");
+    width: 100%;
+    height: 100%;
+    background-color: black;
+    background-position: top;
+    background-repeat: no-repeat;
+    display: flex;
+    /* padding-top: 100%; */
+    /* padding-bottom: 0%; */
+    background-size: cover; /* Faz a imagem cobrir toda a área */
+    flex-direction: column;
+    /* justify-content: center; */
+    align-items: center;
+    position: relative;
+    min-height: 150vh;
+}
+
+.containerTexto {
+    color: white; /* Cor do texto */
+    padding: 20px;
+    margin-top: 18%;
+    border-radius: 8px;
+    max-width: 800px;
+    text-align:justify;
+    z-index: 1; /* Garante que o texto esteja sobre a imagem */
+    position: relative;
+    margin-right: 25%;
+    display: grid;
+    gap: 1rem;
+    font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
+}
+
+.containerTexto h1, .containerTexto p {
+    margin: 0;
+}
+
+.containerTexto h1{
+    text-align: center;
+    gap: 1rem;
+}
+
+.containerTexto p {
+    font-size: large;
+}
