+++ src/app/globals.css
@@ -95,4 +95,8 @@ a {
   text-decoration: none;
 }
 
+p {
+  margin-bottom: 0 !important;
+}
+
