+++ src/app/page.module.css
@@ -9,161 +9,8 @@
   display: flex;
   width: 100%;
   height: 100%;
-  padding: 1rem 1rem;
+  padding: 0 1rem;
   align-items: center;
   justify-content: center;
 }
 
-.containerContato{
-  background-image: url("../../public/bgContato.png");
-	width: 100%;
-  height: 100%;
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
-  display: flex;
-  padding: 3rem 4rem;
-}
-
-.containerSobre {
-  background-image: url("../../public/fundoSobre.png");
-  width: 100%;
-  height: 100%;
-  background-color: black;
-  background-position: top;
-  background-repeat: no-repeat;
-  display: flex;
-  /* padding-top: 100%; */
-  /* padding-bottom: 0%; */
-  background-size: cover; /* Faz a imagem cobrir toda a área */
-  flex-direction: column;
-  /* justify-content: center; */
-  align-items: center;
-  position: relative;
-  min-height: 150vh;
-}
-
-.containerTexto {
-  color: white; /* Cor do texto */
-  padding: 20px;
-  margin-top: 18%;
-  border-radius: 8px;
-  max-width: 800px;
-  text-align:justify;
-  z-index: 1; /* Garante que o texto esteja sobre a imagem */
-  position: relative;
-  margin-right: 25%;
-  display: grid;
-  gap: 1rem;
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
-}
-
-.containerTexto h1, .containerTexto p {
-  margin: 0;
-}
-
-.containerTexto h1{
-  text-align: center;
-  gap: 1rem;
-}
-
-.containerTexto p {
-  font-size: large;
-}
-
-.containerForm{
-  background-color: #fff;
-  width: 50%;
-  height: 100%;
-  border-radius: 8px;
-  -webkit-box-shadow: 0px 2px 16px 1px rgba(0,0,0,0.75);
-  -moz-box-shadow: 0px 2px 16px 1px rgba(0,0,0,0.75);
-  box-shadow: 0px 2px 16px 1px rgba(0,0,0,0.75);
-  padding: 2rem;
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.imgContato{
-  display: flex;
-  position: relative;
-  bottom: 20px;
-  left: 70px;
-  transform: rotate(-14deg);
-}
-
-.containerHome {
-  width: 100%;
-  min-width: 100%;
-  display: flex;
-  height: 100%;
-  min-height: 70vh;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.containerDownloads {
-  display: grid;
-  grid-template-rows: auto 1fr;
-  width: 100%;
-  height: 100%;
-  text-align: center;
-} 
-
-.pageDownloads {
-  display: grid;
-  grid-template-columns: 1fr 4fr;
-  width: 100%;
-  height: 100%;
-}
-
-.sideBarDownloads {
-  display: flex;
-  justify-content: flex-start;
-  gap: 0.5rem;
-  flex-direction: column;
-  padding: 1rem;
-}
-
-.downloads {
-  display: grid;
-  padding: 1rem;
-  grid-template-rows: auto 1fr;
-}
-
-.topBarDownloads{
-  display: grid;
-  align-items: center;
-  grid-template-columns: 3fr 1fr;
-  gap: 1rem;
-}
-
-.buttonCrescente {
-  display: flex;
-  justify-content: space-between;
-  padding: 0 .5rem 0 1rem;
-  border: 1px solid #cccccc;
-  border-radius: 4px;
-  align-items: center;
-  height: 38px;
-  cursor: pointer;
-  text-align: center;
-}
-
-.buttonCrescente:hover {
-  border: 1px solid #9d2053;
-}
-
-.cardsDownloads{
-  display: grid;
-  width: 100%;
-  height: 100%;
-  grid-template-columns: repeat(3, minmax(0, 1fr));
-  padding: 1rem 0;
-  gap: 1rem;
-  min-height: 80vh;
-}
